Integration with BCGov repo workflow
