+++ verifier/.gitignore
@@ -1,8 +1,3 @@
-# /*
-# * 
-# * SPDX-License-Identifier: LiLiQ-R-v.1.1
-# * License-Filename: /LICENSE
-# */
 # 
 #  .gitignore file pour projets ReactJS et ReactNativeJS 
 #
