+++ verifier/voyages-controller/Dockerfile.dev
@@ -7,10 +7,8 @@ COPY public public/
 COPY src /app/src
 
 RUN npm install --silent
-RUN npm install react-scripts@3.4.1 --silent
 RUN npm run build
 
 COPY server.js /app/server.js
-COPY .env /app/.env
 
 EXPOSE 10000
