+++ verifier/voyages-controller/package.json
@@ -20,8 +20,7 @@
     "react-router-hash-link": "^2.2.0",
     "react-scripts": "^3.4.3",
     "reactstrap": "^8.4.1",
-    "serve": "^11.3.2",
-    "uuid": "^8.3.0"
+    "serve": "^11.3.2"
   },
   "scripts": {
     "start": "node server.js",
