+++ verifier/voyages-controller/public/locales/en/vaccine.json
@@ -1,5 +1,5 @@
 {
-    "vaccin": "Vacine", 
+    "vaccin": "Vaccine", 
     "key": "value", 
     "welcomeTitle": "Online ticket reservation ",
     "welcomeMessage": "Your one-stop travel site for your dream vacation. Bundle your stay with a car rental or flight and you can save more.",
@@ -38,8 +38,10 @@
     "vaccine_verified": "Vaccination Proof has been successfully verified", 
     "vaccine_message" : "", 
     "vaccineProof" : "Vaccination Proof Validation", 
+    "lblVaccineProof" : "Vaccination Proof",
     "vaccineQRMsg" : "Please scan this QR code to validate your vaccination proof", 
     "btnVaccineQR": "Validate vaccination proof", 
+    "lblRecipient": "Recipient", 
     "travelCarefreeTitle": "Travel Carefree with Vaccination Certificates!",
     "travelCarefreeSubTitle": "Vaccination certificates made reliable and easy.",
     "travelCarefreeDescription": "This demo shows how an institution can verify a digital immunity certificates of a person either online or face-2-face.",
