+++ verifier/voyages-controller/public/locales/fr/vaccine.json
@@ -38,8 +38,10 @@
     "vaccine_verified": "La Preuve Vaccinale du passager a été vérifiée.", 
     "vaccine_message" : "", 
     "vaccineProof" : "Validation de la Preuve Vaccinale", 
+    "lblVaccineProof" : "Preuve vaccinale",
     "vaccineQRMsg" : "Merci de balayer ce QRCode pour valider votre attestation de preuve vaccinale",
     "btnVaccineQR": "Valider preuve vaccinale", 
+    "lblRecipient": "Bénéficiaire", 
     "travelCarefreeTitle": "Voyagez sans souci avec vos preuves vaccinales",
     "travelCarefreeSubTitle": "Les preuves vaccinales sont fiables et faciles à établir.",
     "travelCarefreeDescription": "Cette démo montre comment une institution peut vérifier les preuves vaccinales d'une personne, en ligne ou en face à face.",
