+++ verifier/voyages-controller/src/VaccineProof/containers/QRVerificationContainer.js
@@ -3,27 +3,29 @@ import { Container, Button, Col, Spinner }  from 'reactstrap'
 import { useTranslation } 					from 'react-i18next' 
 import QRProofComponent                     from '../components/QRProofComponent'
 import { GET_API_SECRET }                   from '../../config/constants'
-import { GET_CRED_DEF_ID } 					from '../../config/constants'
 import { fetchWithTimeout }                 from '../../helpers/fetchWithTimeout'
+import { GET_SCHEMA_NAME }                  from '../../config/constants'
+import { GET_SCHEMA_VERSION }               from '../../config/constants'
 import                                           '../../assets/styles/LoginContainer.css'
 
 
 function QRVerificationContainer(props){
 
 	const { t } = useTranslation(['translation', 'vaccine']); 
-	const [showAuthButton, setAuthButton]  = useState(false);
-	const [showLoader, setLoader]          = useState(false);
 
 	let INTERVAL    = 5000; 
 	let TIMEOUT     = 3000; 
-	let cred_def_id = GET_CRED_DEF_ID();
+	/*let schemaName = GET_SCHEMA_NAME();  
+	let version = GET_SCHEMA_VERSION(); */
+	let schemaName = "vaccine"; 
+	let version ="1.2";
 
     useEffect(() => {
         getConnectionInfo()
     }, []);
 
     function getConnectionInfo() {
-	
+		
 		try {
 			fetchWithTimeout(`/connections/${props.location.state.invitation.connection_id}`,
 				{
@@ -56,12 +58,11 @@ function QRVerificationContainer(props){
     
     function clearIntervalFunction(intervalFunction) {
 		clearInterval(intervalFunction);
-		setAuthButton(true);
+		requestProof()
     }
     
 
     function requestProof(){
-		
 		fetch(`/present-proof/send-request`, 
 			{
 				method : 'POST', 
@@ -82,33 +83,29 @@ function QRVerificationContainer(props){
 							"description": {
 								"name": "description",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"expirationDate": {
 								"name": "expirationDate",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"countryOfVaccination": {
 								"name": "countryOfVaccination",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"credential_type": {
 								"name": "credential_type",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							
@@ -116,96 +113,78 @@ function QRVerificationContainer(props){
 							"recipient_birthDate": {
 								"name": "recipient_birthDate",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"recipient_fullName": {
 								"name": "recipient_fullName",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"recipient_type": {
 								"name": "recipient_type",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							}, 
 							"vaccine_dateOfVaccination": {
 								"name": "vaccine_dateOfVaccination",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							}, 
 							"vaccine_disease": {
 								"name": "vaccine_disease",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"vaccine_type": {
 								"name": "vaccine_type",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
                             }, 
                             "vaccine_medicinalProductName": {
 								"name": "vaccine_medicinalProductName",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							},
 							"vaccine_marketingAuthorizationHolder": {
 								"name": "vaccine_marketingAuthorizationHolder",
 								"restrictions": [
-									{
-									"cred_def_id": cred_def_id
-									}
+									{"schema_name": schemaName,
+                        			"schema_version": version}
 								]
 							}
 						}, 
 						"requested_predicates" : {}
 					}
-				}	
+				}
 				)}).then(response => response.json())
 					.then(data => {
 						props.history.push('/proofResult', {
 							presentation_exchange_id: data.presentation_exchange_id,
 							connection_id           : props.location.state.invitation.connection_id,
-							ticket					: props.location.state.ticket,
 						}
                     );
 				});
 	}
-
-	const handleAuthorisation = () => {
-		setLoader(true);
-		requestProof(); 
-    }
     
     return(
         <div className="Root" style={{ backgroundColor: '#FCF8F7', display: "flex" }}>
 			<Container >
 				<Col>
 					<QRProofComponent value={JSON.stringify(props.location.state)} />
-				</Col>
-				<Col className="mt-3">
-					{showAuthButton && !showLoader ?
-						<Button outline color="primary" onClick={handleAuthorisation}>{t('vaccine:btnVaccineQR')}</Button> : showLoader ? <Spinner /> : null}
 				</Col>
 			</Container>
 		</div>
